Dashboard/requirements.txt
--no-binary :all: pyodbc
python-dotenv
altair
streamlit
pandas
ipykernel